feat: remove deprecated methods and functions
