+++ document.go
@@ -5,7 +5,6 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"log"
 	"slices"
 	"sort"
 	"strings"
@@ -452,23 +451,6 @@ type PrimaryData interface {
 	// If the data node is set to "null" (a jsonapi.One instance with a nil value) then nil is returned.
 	// If the data node itself is nil, First() panics.
 	First() *Resource
-}
-
-// First returns the first item in a primary data node -- the node itself for single or "one"
-// primary data, or the first element in multi or "many" primary data.
-//
-// If the data node is set to "null" (a jsonapi.One instance with a nil value) then nil is returned.
-// If the data node itself is nil, First() panics.
-//
-// Deprecated: First is deprecated and will be removed in the next major release. Use the First()
-// method on the PrimaryData interface instead.
-func First(data PrimaryData) *Resource {
-	log.Println("The First() function is deprecated and will be removed in the next major release. Use the PrimaryData.First() method instead.")
-	items := data.Items()
-	if len(items) == 0 {
-		return nil
-	}
-	return items[0]
 }
 
 // IsMany returns false, as this is a "to-one" relationship.
